requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
django-generate-secret-key==1.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
mysqlclient==2.0.3
pycodestyle==2.7.0
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.60.0
urllib3==1.26.5
